refactor: LInkedList 함수 생성
* LInkedList 함수 생성 후 기존의 CalculatorItemQueue의 구현을 옮김
* CalculatorItemQueue에서는 LinkedList의 함수 메서드 불러와서 재사용
