+++ Calculator/Calculator/Model/CalculatorItemQueue.swift
@@ -8,49 +8,21 @@
 import Foundation
 
 struct CalculatorItemQueue<T>: CalculateItem {
-
-    var front: Node<T>?
-    
-    init(front: Node<T>? = nil) {
-        self.front = front
-    }
+    var queue = LinkedList<T>()
     
     mutating func enQueue(_ data: T?) {
-        var node = front
-        
-        if node == nil {
-            front = Node(data: data)
-            return
-        }
-        
-        while node?.next != nil {
-            node = node?.next
-        }
-        node?.next = Node(data: data)
+        queue.append(data)
     }
     
     mutating func deQueue() {
-        if front == nil {
-            return
-        }
-        front = front?.next
+        queue.deQueue()
     }
     
     mutating func removeAll() {
-        front = nil
+        queue.removeAll()
     }
     
     mutating func returnQueue() -> Array<T> {
-        var queue: Array<T> = []
-        var node = front
-        
-        while node?.data != nil {
-            guard let data = node?.data else {
-                return []
-            }
-            queue.append(data)
-            node = node?.next
-        }
-        return queue
+        return queue.returnQueue()
     }
 }
